% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{caf_nap_recall}
\alias{caf_nap_recall}
\title{Caffeine, Naps, Placebos}
\format{A \code{data.frame} consisting of \strong{61} rows and \strong{2} columns.

Variables included are:
\itemize{
\item \code{words_recalled}: Number of words recalled after 20 mins in the
verbal task described in the study.
\item \code{treatment}: One of \code{placebo}, \code{nap} (a 60 - 90 minute nap), or
\code{caffeine} (a 200 mg pill).
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2603066/}
}
\usage{
caf_nap_recall
}
\description{
Each row is a "subject" in the experiment described in
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2603066/}{Comparing the benefits of Caffeine, Naps and Placebo on Verbal, Motor and Perceptual Memory}.
We say "subject" because we are simulating data to best match the results on
this study as the data is not publicly available.
}
\keyword{datasets}
