% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalpiaz-package.R
\docType{package}
\name{dalpiaz-package}
\alias{dalpiaz}
\alias{dalpiaz-package}
\title{dalpiaz: Dalpiaz Misc}
\description{
Provides a collection of data sets and analysis functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daviddalpiaz/dalpiaz}
  \item Report bugs at \url{https://github.com/daviddalpiaz/dalpiaz/issues}
}

}
\author{
\strong{Maintainer}: David Dalpiaz \email{dalpiaz2@illinois.edu}

Authors:
\itemize{
  \item James Balamuta \email{balamut2@illinois.edu}
}

}
\keyword{internal}
