% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ncaa_2019_mens_march_madness}
\alias{ncaa_2019_mens_march_madness}
\title{2019 NCAA Men's March Madness}
\format{A \code{data.frame} consisting of \strong{33} rows and \strong{10} columns.

Variables included are:
\itemize{
\item \code{highseed}: Seed of the "higher" ranked team.
\item \code{highseedschool}: School of the "higher" ranked team.
\item \code{highseedscore}: Score of the "higher" ranked team.
\item \code{lowseed}: Seed of the "lower" ranked team.
\item \code{lowseedschool}: School of the "lower" ranked team.
\item \code{lowseedscore}: Score of the "lower" ranked team.
\item \code{region}: Region the game was played in. One of: \code{East}, \code{Midwest}, \code{West}, \code{South}
\item \code{date}: Date the game was played.
\item \code{round}: Round in which the game was played. One of : \code{64}, \code{32}, \code{16}, \code{8}, \code{4}, \code{2}
}}
\source{
\url{https://en.wikipedia.org/wiki/2019_NCAA_Division_I_Men\%27s_Basketball_Tournament}
}
\usage{
ncaa_2019_mens_march_madness
}
\description{
Each row is a game played in the
\href{https://en.wikipedia.org/wiki/2019_NCAA_Division_I_Men\%27s_Basketball_Tournament}{2019 NCAA Division I Men's Basketball Tournament}.
}
\keyword{datasets}
